backend/requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
wtforms
psycopg2-binary
requests
cryptography
polars
scikit-learn
numpy
matplotlib
tensorflow
keras
nltk